ARM: dts: k2g: disable dma-coherency for PCIe
The DMA coherency is broken for Keystone PCIe in the latest kernel . Add
an intermediate solution of removing the dma-coherency property in
the DT bindings to the driver functional for PCI network devices. Currently
because of this issue, the PCIe NIC cards doesn't function. This used to
work before. Even with this removed, we get reasonable performance with
intel e1000e PCI card comparable to other TI devices. So for now add this
intermediate solution before the coherency issue is root caused. This
appears to be a problem with the dma-coherency mask set incorrectly for
the PCI network device.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>